Prueba de Solucion Categoria con base de datos exitosa/feat: refactorización del módulo de categorías y correcciones en GenericService
