+++ MindPulse.Core.Application/DTOs/Categories/CategoryDTO.cs
@@ -4,5 +4,7 @@ namespace MindPulse.Core.Application.DTOs.Categories
     {
         public int Id { get; set; } 
         public string Name { get; set; }
+        public string Description { get; set; }
+
     }
 }
