+++ MindPulse.Core.Application/DTOs/Recommendations/RecommendationDTO.cs
@@ -1,0 +1,13 @@
+namespace MindPulse.Core.Application.DTOs.Recommendations
+{
+    public class RecommendationDTO
+    {
+        public int Id { get; set; }
+        public string Title { get; set; }
+        public string Content { get; set; }
+        public string? UrlSource { get; set; }
+
+        public int UserId { get; set; }
+        public int CategoryId { get; set; }
+    }
+}
