+++ MindPulse.Core.Application/Interfaces/Services/IRecommendations/IEducationalContentService.cs
@@ -1,0 +1,9 @@
+using MindPulse.Core.Application.DTOs.Recommendations;
+using MindPulse.Core.Domain.Entities.Recommendations;
+
+namespace MindPulse.Core.Application.Interfaces.Services.Recommendations
+{
+    public interface IEducationalContentService : IGenericService<EducationalContentDTO, EducationalContentDTO, EducationalContent, EducationalContentDTO>
+    {
+    }
+}
