+++ MindPulse.Core.Domain/Entities/Categories/Category.cs
@@ -1,12 +1,15 @@
 using MindPulse.Core.Domain.Commons;
-using MindPulse.Core.Domain.Entities.Evaluations;
+using MindPulse.Core.Domain.Entities.Recommendations;
 
 namespace MindPulse.Core.Domain.Entities.Categories
 {
     public class Category : BaseEntity
     {
         public string Name { get; set; }
+        public string Description { get; set; }
 
-        public ICollection<Test> Evaluations { get; set; }
+        // Relaciones
+        public ICollection<Recommendation> Recommendations { get; set; }
+        public ICollection<EducationalContent> EducationalContents { get; set; }
     }
 }
