+++ MindPulse.WebApp/appsettings.json
@@ -1,34 +1,29 @@
 {
-  "ConnectionStrings": {
-    "DefaultConnection": "Server=MYKE\\SQLEXPRESS;Database=MindPulseApp;Trusted_Connection=True;MultipleActiveResultSets=true;TrustServerCertificate=True;"
-  },
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
+    "ConnectionStrings": {
+        "DefaultConnection": "Server=BRANITRO\\SQLEXPRESS;Database=MindPulseApp;Trusted_Connection=True;MultipleActiveResultSets=true;TrustServerCertificate=True;"
+    },
+    "Logging": {
+        "LogLevel": {
+            "Default": "Information",
+            "Microsoft.AspNetCore": "Warning"
+        }
+    },
+    "AllowedHosts": "*",
+    "EmailSettings": {
+        "EmailFrom": "mindpulseproject@gmail.com",
+        "SmtpHost": "smtp.gmail.com",
+        "SmtpPort": 587,
+        "SmtpUser": "mindpulseproject@gmail.com",
+        "SmtpPass": "oijqfwikeegssfld",
+        "DisplayName": "MindPulse"
+    },
+    "AppSettings": {
+        "BackendUrl": "https://localhost:7221/"
+    },
+    "Jwt": {
+        "Key": "aVerySecureRandomGeneratedKey12345678901234567890",
+        "Issuer": "MindPulse",
+        "Audience": "MindPulseUsers",
+        "DurationInMinutes": 60
     }
-  },
-  "AllowedHosts": "*",
-
-  "EmailSettings": {
-    "EmailFrom": "mindpulseproject@gmail.com",
-    "SmtpHost": "smtp.gmail.com",
-    "SmtpPort": 587,
-    "SmtpUser": "mindpulseproject@gmail.com",
-    "SmtpPass": "oijqfwikeegssfld",
-    "DisplayName": "MindPulse"
-  },
-
-
-  "AppSettings": {
-    "BackendUrl": "https://localhost:7221/"
-  },
-
-  "Jwt": {
-    "Key": "aVerySecureRandomGeneratedKey12345678901234567890",
-    "Issuer": "MindPulse",
-    "Audience": "MindPulseUsers",
-    "DurationInMinutes": 60
-  }
 }
-
